requirements.txt
fastapi
uvicorn[standard]

python-dotenv
discord.py

gspread
oauth2client

pydantic-settings